delete files through a button
now you don't have to add the b=true query parameter to delete files.

+++ file_view.php
@@ -6,29 +6,51 @@
     $item = '';
     $dir = "./uploads/$name/";
     if(isset($_GET['item'])){
+        $_SESSION['curr_item'] = $_GET['item'];
         $item = $_GET['item'];
     }
     $file_dir = $dir.$item;
 
-    if(isset($_GET['b']) && $_GET['b'] == true){
-        if(file_exists($file_dir)){
-            $files = array_diff( scandir($dir), array('.','..') );
-            $item = explode(".", $item)[0];
+    if( $_SERVER['REQUEST_METHOD'] === "POST" ){
+        if(isset($_POST['delete'])){
 
-
-            foreach( $files as $file ){
-                $compare = explode(".", $file)[0];
-                if($item === $compare){
-                    unlink($dir.$file);
-                }
-            }
-            $exists = false;
-            header("Location: ./files.php?m=true");
-            exit();
-        }else{
-            echo "File does not exist...";
+            $item = $_SESSION['curr_item'];
+            header("Location: ".$_SERVER['PHP_SELF']."?item=$item&b=true");
+            
         }
     }
+
+    if( $_SERVER['REQUEST_METHOD'] === "GET" ){
+
+        if(isset($_GET['b']) && $_GET['b'] == true){
+            if(file_exists($file_dir)){
+                $files = array_diff( scandir($dir), array('.','..') );
+                $item = explode(".", $item)[0];
+    
+                foreach( $files as $file ){
+                    $compare = explode(".", $file)[0];
+                    if($item === $compare){
+                        unlink($dir.$file);
+                        $exists = false;
+                    }
+                }
+                header("Location: ./files.php?m=true");
+                exit();
+            }else{
+                echo "File does not exist...";
+            }
+        }
+
+    }
+
+    function addOrUpdateUrlParam($name, $value)
+    {
+        $params = $_GET;
+        unset($params[$name]);
+        $params[$name] = $value;
+        return basename($_SERVER['PHP_SELF']).'?'.http_build_query($params);
+    }
+
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -53,6 +75,9 @@
             }
         ?>
     </p>
+    <form action="<?php echo $_SERVER['PHP_SELF'] ?>" method="post">
+        <button type="submit" name="delete" class="btn btn-primary">Delete Text and Image</button>
+    </form>
     <script src="./index.js"></script>
 </body>
 </html>
